+++ packages/ethers/lib/_version.d.ts
@@ -1,2 +1,2 @@
-export declare const version = "ethers/5.0.27";
+export declare const version = "ethers/5.0.28";
 //# sourceMappingURL=_version.d.ts.map