+++ packages/providers/lib/pocket-provider.d.ts
@@ -1,9 +1,11 @@
-import { Network } from "@ethersproject/networks";
+import { Network, Networkish } from "@ethersproject/networks";
 import { ConnectionInfo } from "@ethersproject/web";
 import { UrlJsonRpcProvider } from "./url-json-rpc-provider";
 export declare class PocketProvider extends UrlJsonRpcProvider {
     readonly applicationId: string;
     readonly applicationSecretKey: string;
+    readonly loadBalancer: boolean;
+    constructor(network?: Networkish, apiKey?: any);
     static getApiKey(apiKey: any): any;
     static getUrl(network: Network, apiKey: any): ConnectionInfo;
     isCommunityResource(): boolean;